Updated the implementation of the GetCultureInfo extension method
With Umbraco 8.1, we can now use the GetCultureFromDomains extension method for determining the culture of a given node. This method will however return the culture code as a string - opposed to returning an instance of CultureInfo. Our own extension method is therefor responsible for getting the corresponding CultureInfo instance. The result of CultureInfo.GetCultureInfo is cached internally in .NET making sure we get the same instance when calling the method twice with the same culture code.