Replaced SpaControllerBase class with new SpaRequestHelper class
The idea was that in client projects, we would inherit from the SpaControllerBase, and than override the necessary methods. But in some cases, we need to inherit from a client specific controller instead, which wasn't possible without lacking the methods of SpaControllerBase.

With this commit, SpaControllerBase is now gone, the same logic has now been moved to a new SpaRequestHelper class