GetData endpoint should always return 200 OK
To make the API easier to work with for a frontenders, ann non-error responses should return a 200 OK instead of whatever status code it should have been. In the JSON response, the actual status code (eg. 301, 307 or 404) is then available through "meta.code". This change in behavior thereby fixes #12

This is a breaking change, so if we're upgrading our older solutions, the "SpaRequestHelper.OverwriteStatusCodes" property can be set to false to disable this behavior.

The SPA package also has problems with the correct status code being lost for cached models, as the status isn't set on the cached "SpaDataModel", but rather on the "SpaRequest", which isn't cached. With this commit, the "SpaDataModel" class now exposes a "Meta" property, which persists the status code even when being cached. This thereby also fixes #13