Updated micro caching layer to set the correct culture info
When requests are handled by the SPA package, the culture info is correctly set on the SPA request model - but when using our micro caching layer (typically only on stage and live), the culture info of the request that was cached is lost. With this commit, we're now caching a "SpaCachedModel" instead of the "SpaDataModel". This allows us to save a bit more information that is then cached ... eg. the culture info.