Introduced new "ISpaContentModel" interface
In the past, the content model of a "SpaRequest" must have been an instance of "SpaContentModel", which extends Umbraco's "PublishedContentModel" class (which typically reflects a ModelsBuilder generated model). This has worked well as we historically have returned the ModelsBuilder models directly in the API.

However for some of our newer sites (e.g. our Citi sites), we instead have tailored view models, that we're returning instead. To support this, introducing the "ISpaContentModel" interface was necessary.